feature: storage screen
